+++ tests/unit/helpers/FileHelper/FileHelperTest.php
@@ -154,6 +154,17 @@ class FileHelperTest extends Unit
     }
 
     /**
+     * @dataProvider getExtensionByMimeTypeDataProvider
+     *
+     * @param string $expected
+     * @param string $mimeType
+     */
+    public function testGetExtensionByMimeType(string $expected, string $mimeType)
+    {
+        self::assertSame($expected, FileHelper::getExtensionByMimeType($mimeType));
+    }
+
+    /**
      * @dataProvider sanitizeFilenameDataProvider
      *
      * @param string $expected
@@ -338,6 +349,17 @@ class FileHelperTest extends Unit
             [true, __DIR__ . '/sandbox/isdirempty/yes'],
             [false, __DIR__ . '/sandbox/isdirempty/no'],
             [false, __DIR__ . '/sandbox/isdirempty/dotfile'],
+        ];
+    }
+
+    /**
+     * @return array
+     */
+    public function getExtensionByMimeTypeDataProvider(): array
+    {
+        return [
+            ['jpg', 'image/jpeg'],
+            ['svg', 'image/svg+xml'],
         ];
     }